migrate connection types to be with other messaging
